requirements.txt
lldb-python==19.0.0.dev1
pytest==8.3.3
